+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v16.13.6] - 2024-08-03
+
+### Fixed
+- Improving ContentSecurityPolicy with removing unsafe-inline from script-src
+
 ## [v16.13.5] - 2024-08-03
 
 ### Fixed